+++ src/renderer/src/components/multi-player/src/core/index.ts
@@ -1,8 +1,4 @@
 import ArtPlayerAdapter from './artplayer';
-import DPlayerAdapter from './dplayer';
-import NPlayerAdapter from './nplayer';
-import OPlayerAdapter from './oplayer';
-import VlcPlayerAdapter from './vlcplayer';
 import XgPlayerAdapter from './xgplayer';
 
-export { ArtPlayerAdapter, DPlayerAdapter, NPlayerAdapter, OPlayerAdapter, VlcPlayerAdapter, XgPlayerAdapter };
+export { ArtPlayerAdapter, XgPlayerAdapter };