+++ src/renderer/src/components/multi-player/src/core/nplayer/plugins/index.ts
@@ -1,4 +1,0 @@
-import pipPlugin from './pip';
-import playNextPlugin from './playNext';
-
-export { pipPlugin, playNextPlugin };