+++ src/renderer/src/components/multi-player/src/core/oplayer/types.ts
@@ -1,8 +1,0 @@
-import type { Player } from '@oplayer/core';
-
-export interface PlayerAdapter extends Player {
-  storage?: {
-    set: (key: string, value: any) => void;
-    get: (key: string) => any;
-  };
-}