+++ src/shared/locales/lang/zh-CN/pages/md.ts
@@ -7,24 +7,16 @@ const binaryHelpMD = `- 点击安装会添加到任务队列, 无需多次点击
 
 const customPlayerMD = `> 该功能依赖于系统命令, 可调用\`环境变量\`或\`指定路径\`
 
-- 播放器类型: Vlc播放器
-  - 该模式做了专门输配
-  - 如何设置: \`libvlc so/lib/dylib路径\`;\`libvlc plugin路径\`
-  - 常见
-    - Windows: \`C:\\Program Files (x86)\\VideoLAN\\VLC\\libvlc.dll;C:\\Program Files (x86)\\VideoLAN\\VLC\\plugins\`
-    - Mac: \`/Applications/VLC.app/Contents/MacOS/lib/libvlc.dylib;/Applications/VLC.app/Contents/MacOS/plugins\`
-    - Linux: \`/usr/lib/x86_64-linux-gnu/libvlc.so;/usr/lib/vlc/plugins\`
-- 播放器类型: 自定义播放器
-  - 如何设置:
-    - Windows: \`[右键图标->右键属性->目标]\`查看路径
-    - Mac: \`/Applications/软件名.app\`
-    - Linux: \`[whereis 软件名]\`查看路径
-  - 常见播放器\`[根据实际修改路径]\`
-    - PotPlayer(Window): \`C:\\Program Files (x86)\\PotPlayer\\PotPlayerMini64.exe\`
-    - VLC(Window): \`C:\\Program Files (x86)\\VideoLAN\\VLC\\vlc.exe\`
-    - IINA(Mac): \`/Applications/IINA.app\`
-    - VLC(Mac): \`/Applications/VLC.app\`
-    - VLC(Linux): \`/usr/bin/vlc\`
+- 如何设置:
+  - Windows: \`[右键图标->右键属性->目标]\`查看路径
+  - Mac: \`/Applications/软件名.app\`
+  - Linux: \`[whereis 软件名]\`查看路径
+- 常见播放器\`[根据实际修改路径]\`
+  - PotPlayer(Window): \`C:\\Program Files (x86)\\PotPlayer\\PotPlayerMini64.exe\`
+  - VLC(Window): \`C:\\Program Files (x86)\\VideoLAN\\VLC\\vlc.exe\`
+  - IINA(Mac): \`/Applications/IINA.app\`
+  - VLC(Mac): \`/Applications/VLC.app\`
+  - VLC(Linux): \`/usr/bin/vlc\`
 `;
 
 const disclaimerMD = `> 发布日期: 2026年4月30日